Remove the unit info in the collada file.
This info is not used in fuerte, but used in groovy.
As a result, in groovy, the size of the bodies is
reduced by 100.
Since the dimensions are in meters in the dae files
owing to an additional scaling processs, the unit info
was set to meters, thus ensuring fuerte and
groovy compatibility

See http://answers.gazebosim.org/question/581/is-gazebo-125-unit-aware/
and http://answers.ros.org/question/59514/did-rviz-become-unit-aware/